requirements.txt
pandas
openpyxl
python-dotenv
newsapi-python
pypdf
python-multipart
google-generativeai
